requirements.txt
beautifulsoup4==4.13.3
bs4==0.0.2
certifi==2025.1.31
chardet==3.0.4
charset-normalizer==3.4.1
feedparser==6.0.11
idna==2.8
python-dateutil==2.9.0.post0
requests==2.32.3
sgmllib3k==1.0.0
six==1.17.0
soupsieve==2.6
typing_extensions==4.12.2
urllib3==2.3.0
